Pin all third party GitHub actions to a specific SHA commit
Previously, the third party GitHub actions were set to a git tag, but
git tags are mutable. This means that a malicious agent could switch the
tag out from under us to inject new malicious code. However, a SHA is
immutable. Once set in the git repository the SHA will never refer a
different instance of the code. This is security improvement.

Using a SHA is considered best practices in GitHub actions and discussed
in the official docs at:


https://docs.github.com/en/actions/reference/security/secure-use#using-third-party-actions

> You can help mitigate this risk by following these good practices:
>
> - Pin actions to a full-length commit SHA